move empennage models to separate file
